storage_types: extract Mutation
