+++ components/storage_types/Cargo.toml
@@ -3,3 +3,6 @@ name = "storage_types"
 version = "0.1.0"
 edition = "2018"
 publish = false
+
+[dependencies]
+keys = { path = "../keys" }
